// src/asesor-gerencia/entities/asesor-gerencia-estado.enum.ts
// src/asesor-gerencia/entities/asesor-gerencia-estado.enum.ts
export enum AsesorGerenciaEstado {
  DISPONIBLE = 'DISPONIBLE',
  EN_REVISION = 'EN_REVISION',
  COMPLETADO_ASESOR_GERENCIA = 'COMPLETADO_ASESOR_GERENCIA',
  OBSERVADO_ASESOR_GERENCIA = 'OBSERVADO_ASESOR_GERENCIA',
  RECHAZADO_ASESOR_GERENCIA = 'RECHAZADO_ASESOR_GERENCIA',
}